Added a check for stale classes (classes open for more than 9 hours)
Specifics are as follows:
Classes open more than 9 hours are automatically closed.
If the class has students who have signed out the classes end time will
be adjusted to the final students departure time, or the final
underrated students start time.
If the class has students who did not sign out their time will be
adjusted to the last departing student time or the last sign in time,
whichever is most recent.
If the class had no students the close time will be adjusted to the
open time, negating the class.
If the class had students but none signed out, all departure times for
students and class close time will be set to the last students signing
time.